% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscraper.R
\docType{data}
\name{denverPropertyDF}
\alias{denverPropertyDF}
\title{Scrapes denver property website for data for a single parcel}
\format{An object of class \code{data.frame} with 1 rows and 33 columns.}
\usage{
denverPropertyDF
}
\arguments{
\item{parcelID}{A string containing the parcel ID for the property}

\item{rowNumber}{A number containing the row number for this property parcel}
}
\description{
Scrapes denver property website for data for a single parcel
}
\keyword{datasets}
