% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLocalData.R
\docType{data}
\name{localData}
\alias{localData}
\title{read local data into Rds objects}
\format{An object of class \code{data.frame} with 208420 rows and 46 columns.}
\usage{
localData
}
\value{
Two objects: a) a dataframe holding residential characteristics for all 
properties, and b) a vector containing strings of all the property parcel ids for denver
}
\description{
read local data into Rds objects
}
\keyword{datasets}
