% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscraperCollect.R
\docType{data}
\name{parcelVector}
\alias{parcelVector}
\title{Iteratevely runs webscraper function through vector of parcel ids}
\format{An object of class \code{character} of length 208420.}
\usage{
parcelVector
}
\description{
Iteratevely runs webscraper function through vector of parcel ids
}
\keyword{datasets}
