% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscraper.R
\name{webscraper}
\alias{webscraper}
\title{Scrapes denver property website for data for a single parcel}
\usage{
webscraper()
}
\arguments{
\item{parcelID}{A string containing the parcel ID for the property}

\item{rowNumber}{A number containing the row number for this property parcel}
}
\value{
A list object containing data for a single property
}
\description{
Scrapes denver property website for data for a single parcel
}
\examples{
webscraper("0503111031031", 1) 

}
